extract Post struct in a separate file
